desarrollar funcionalidad para consumir api /api/items/offset

+++ Spacehub.Application/Dtos/ItemDto.cs
@@ -5,7 +5,7 @@ public class ItemDto
   public int IdItem { get; set; }
   public required string ItemName { get; set; }
   public required string ItemImage { get; set; }
-  public double ItemUnitPrice { get; set; }
+  public decimal ItemUnitPrice { get; set; }
 }
 
 public class ItemDetailDto
@@ -14,7 +14,7 @@ public class ItemDetailDto
   public required string ItemName { get; set; }
   public required string ItemDescription { get; set; }
   public required string ItemImage { get; set; }
-  public double ItemUnitPrice { get; set; }
+  public decimal ItemUnitPrice { get; set; }
   public int Itemstock { get; set; }
   public required string ItemCategory { get; set; }
 }
