+++ Spacehub.Application/Models/ItemsModel.cs
@@ -11,7 +11,7 @@ public class Item
   public required string item_name { get; set; }
   public required string item_description { get; set; }
   public required string path_image { get; set; }
-  public double unit_price { get; set; }
+  public decimal unit_price { get; set; }
   public int stock { get; set; }
 
   [ForeignKey("id_category")]
